chore: upgrade monorepo to TypeScript 7
Move the default compiler and declaration generation to TypeScript 7 while retaining TypeScript 6 only for tools that require the classic compiler API.

Migrate the remaining tsup packages to tsdown so builds no longer depend on tsup's compiler API integration.
