+++ .changeset/beige-jobs-scream.md
@@ -1,0 +1,5 @@
+---
+'@mastra/editor': patch
+---
+
+Preserved tool schemas when applying stored descriptions.
