+++ .changeset/beige-suits-report.md
@@ -1,0 +1,5 @@
+---
+'@mastra/oracledb': patch
+---
+
+Improved package builds with TypeScript 7.
