+++ .changeset/busy-news-build.md
@@ -1,0 +1,5 @@
+---
+'@mastra/codemod': patch
+---
+
+Improved codemod development compatibility with TypeScript 7.
