+++ .changeset/deep-radios-tap.md
@@ -1,0 +1,5 @@
+---
+'@mastra/telegram': patch
+---
+
+Improved package builds with TypeScript 7.
