+++ .changeset/easy-cities-repeat.md
@@ -1,0 +1,5 @@
+---
+'@mastra/turso': patch
+---
+
+Improved package builds with TypeScript 7.
