+++ .changeset/empty-frogs-tie.md
@@ -1,0 +1,5 @@
+---
+'mastracode': patch
+---
+
+Improved Mastra Code development compatibility with TypeScript 7.
