+++ .changeset/giant-baths-begin.md
@@ -1,0 +1,5 @@
+---
+'@mastra/memory': patch
+---
+
+Improved memory package development compatibility with TypeScript 7.
