+++ .changeset/lemon-bears-work.md
@@ -1,0 +1,5 @@
+---
+'@mastra/server': patch
+---
+
+Improved server type generation with TypeScript 7.
