+++ .changeset/quiet-ideas-say.md
@@ -1,0 +1,5 @@
+---
+'@mastra/factory': patch
+---
+
+Improved Factory development compatibility with TypeScript 7.
