+++ .changeset/shy-waves-fold.md
@@ -1,0 +1,5 @@
+---
+'@mastra/deployer': patch
+---
+
+Preserved project builds while adopting TypeScript 7 tooling.
