+++ .changeset/tame-frogs-go.md
@@ -1,0 +1,5 @@
+---
+'@mastra/code-sdk': patch
+---
+
+Improved code SDK development compatibility with TypeScript 7.
