+++ .changeset/tangy-flowers-care.md
@@ -1,0 +1,5 @@
+---
+'@mastra/deepeval': patch
+---
+
+Improved package builds with TypeScript 7.
