+++ .changeset/upset-icons-yell.md
@@ -1,0 +1,5 @@
+---
+'mastra': patch
+---
+
+Improved CLI type checking with TypeScript 7.
