+++ .changeset/yellow-laws-behave.md
@@ -1,0 +1,5 @@
+---
+'@mastra/agent-builder': patch
+---
+
+Improved Agent Builder development compatibility with TypeScript 7.
