+++ mastracode/factory-ui/package.json
@@ -48,7 +48,7 @@
     "react-grab": "^0.1.37",
     "tailwindcss": "4.2.2",
     "tw-animate-css": "^1.4.0",
-    "typescript": "catalog:",
+    "typescript": "catalog:ts6",
     "vite": "^6.0.0",
     "vitest": "catalog:"
   },
