+++ package.json
@@ -24,6 +24,7 @@
     "tsx": "catalog:",
     "turbo": "^2.10.5",
     "typescript": "catalog:",
+    "typescript-classic": "npm:typescript@^6.0.3",
     "vite": "^7.3.1",
     "vitest": "catalog:",
     "zod": "catalog:"
