+++ packages/_config/package.json
@@ -23,6 +23,7 @@
     "eslint-plugin-testing-library": "^7.16.2",
     "eslint-plugin-unicorn": "^72.0.0",
     "globals": "^16.5.0",
+    "typescript": "catalog:ts6",
     "typescript-eslint": "^8.57.0"
   },
   "devDependencies": {
