+++ packages/deployer/package.json
@@ -145,7 +145,7 @@
     "tsdown": "0.22.9",
     "tsx": "catalog:",
     "type-fest": "^5.4.4",
-    "typescript": "catalog:",
+    "typescript": "catalog:ts6",
     "vitest": "catalog:",
     "zod": "catalog:"
   },
