+++ packages/editor/src/namespaces/agent.ts
@@ -1246,12 +1246,13 @@ export class EditorAgentNamespace extends CrudEditorNamespace<
 
     let nextTools: ToolsInput | undefined;
     for (const [toolKey, toolConfig] of Object.entries(storedTools)) {
-      if (!toolConfig.description || !(toolKey in codeTools)) {
+      const codeTool = codeTools[toolKey];
+      if (!toolConfig.description || !codeTool) {
         continue;
       }
 
       nextTools ??= { ...codeTools };
-      nextTools[toolKey] = { ...codeTools[toolKey], description: toolConfig.description };
+      nextTools[toolKey] = { ...codeTool, description: toolConfig.description };
     }
 
     return nextTools ?? codeTools;
