+++ packages/playground/package.json
@@ -120,7 +120,7 @@
     "tailwindcss": "4.3.3",
     "tw-animate-css": "^1.4.0",
     "msw": "^2.6.0",
-    "typescript": "catalog:",
+    "typescript": "catalog:ts6",
     "vite": "^7.3.1",
     "vitest": "catalog:"
   }
