+++ stores/oracledb/tsconfig.json
@@ -1,5 +1,5 @@
 {
   "extends": "../../tsconfig.node.json",
-  "include": ["src/**/*", "tsup.config.ts"],
+  "include": ["src/**/*", "tsdown.config.ts"],
   "exclude": ["node_modules", "**/*.test.ts"]
 }
