+++ stores/oracledb/turbo.json
@@ -6,7 +6,7 @@
       "dependsOn": ["^build"],
       "inputs": [
         "src/**",
-        "tsup.config.ts",
+        "tsdown.config.ts",
         "tsconfig.json",
         "package.json",
         "!**/*.md",
