+++ stores/turso/tsdown.config.ts
@@ -1,15 +1,14 @@
 import { generateTypes } from '@internal/types-builder';
-import { defineConfig } from 'tsup';
+import { defineConfig } from 'tsdown';
 
 export default defineConfig({
   entry: ['src/index.ts'],
   format: ['esm', 'cjs'],
+  fixedExtension: false,
+  nodeProtocol: 'strip',
   clean: true,
   dts: false,
-  splitting: true,
-  treeshake: {
-    preset: 'smallest',
-  },
+  treeshake: true,
   sourcemap: true,
   onSuccess: async () => {
     await generateTypes(process.cwd());
